+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/service/MapperSupportMaintainServiceImpl.java
@@ -1,27 +1,20 @@
 package com.dwarfeng.fdr.impl.service;
 
-import com.dwarfeng.fdr.impl.handler.MapperSupporter;
 import com.dwarfeng.fdr.stack.bean.entity.MapperSupport;
 import com.dwarfeng.fdr.stack.service.MapperSupportMaintainService;
 import com.dwarfeng.subgrade.impl.service.DaoOnlyEntireLookupService;
 import com.dwarfeng.subgrade.impl.service.DaoOnlyPresetLookupService;
 import com.dwarfeng.subgrade.impl.service.GeneralBatchCrudService;
-import com.dwarfeng.subgrade.sdk.exception.ServiceExceptionHelper;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
 import com.dwarfeng.subgrade.stack.bean.dto.PagedData;
 import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
 import com.dwarfeng.subgrade.stack.exception.ServiceException;
-import com.dwarfeng.subgrade.stack.exception.ServiceExceptionMapper;
-import com.dwarfeng.subgrade.stack.log.LogLevel;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
-import java.util.stream.Collectors;
 
 @Service
 public class MapperSupportMaintainServiceImpl implements MapperSupportMaintainService {
@@ -30,26 +23,14 @@ public class MapperSupportMaintainServiceImpl implements MapperSupportMaintainSe
     private final DaoOnlyEntireLookupService<MapperSupport> entireLookupService;
     private final DaoOnlyPresetLookupService<MapperSupport> presetLookupService;
 
-    private final List<MapperSupporter> mapperSupporters;
-
-    private final ServiceExceptionMapper sem;
-
     public MapperSupportMaintainServiceImpl(
             GeneralBatchCrudService<StringIdKey, MapperSupport> crudService,
             DaoOnlyEntireLookupService<MapperSupport> entireLookupService,
-            DaoOnlyPresetLookupService<MapperSupport> presetLookupService,
-            List<MapperSupporter> mapperSupporters,
-            ServiceExceptionMapper sem
+            DaoOnlyPresetLookupService<MapperSupport> presetLookupService
     ) {
         this.crudService = crudService;
         this.entireLookupService = entireLookupService;
         this.presetLookupService = presetLookupService;
-        if (Objects.isNull(mapperSupporters)) {
-            this.mapperSupporters = new ArrayList<>();
-        } else {
-            this.mapperSupporters = mapperSupporters;
-        }
-        this.sem = sem;
     }
 
     @Override
@@ -275,25 +256,6 @@ public class MapperSupportMaintainServiceImpl implements MapperSupportMaintainSe
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public List<MapperSupport> lookupAsList(String preset, Object[] objs, PagingInfo pagingInfo) throws ServiceException {
         return presetLookupService.lookupAsList(preset, objs, pagingInfo);
-    }
-
-    @Override
-    @BehaviorAnalyse
-    public void reset() throws ServiceException {
-        try {
-            List<StringIdKey> mapperKeys = entireLookupService.lookupAsList().stream()
-                    .map(MapperSupport::getKey).collect(Collectors.toList());
-            crudService.batchDelete(mapperKeys);
-            List<MapperSupport> mapperSupports = mapperSupporters.stream().map(supporter -> new MapperSupport(
-                    new StringIdKey(supporter.provideType()),
-                    supporter.provideLabel(),
-                    supporter.provideDescription(),
-                    supporter.provideExampleParam()
-            )).collect(Collectors.toList());
-            crudService.batchInsert(mapperSupports);
-        } catch (Exception e) {
-            throw ServiceExceptionHelper.logParse("重置映射器支持时发生异常", LogLevel.WARN, e, sem);
-        }
     }
 
     /**