+++ fdr-stack/src/main/java/com/dwarfeng/fdr/stack/service/WasherSupportMaintainService.java
@@ -2,7 +2,6 @@ package com.dwarfeng.fdr.stack.service;
 
 import com.dwarfeng.fdr.stack.bean.entity.WasherSupport;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
-import com.dwarfeng.subgrade.stack.exception.ServiceException;
 import com.dwarfeng.subgrade.stack.service.BatchCrudService;
 import com.dwarfeng.subgrade.stack.service.EntireLookupService;
 import com.dwarfeng.subgrade.stack.service.PresetLookupService;
@@ -18,11 +17,4 @@ public interface WasherSupportMaintainService extends BatchCrudService<StringIdK
 
     String ID_LIKE = "id_like";
     String LABEL_LIKE = "label_like";
-
-    /**
-     * 重置清洗器支持。
-     *
-     * @throws ServiceException 服务异常。
-     */
-    void reset() throws ServiceException;
 }