+++ CHANGELOG.md
@@ -34,7 +34,11 @@
 
 ### 功能移除
 
-- (无)
+- 去除支持实体维护服务的重置功能。
+  - com.dwarfeng.fdr.stack.service.FilterSupportMaintainService。
+  - com.dwarfeng.fdr.stack.service.MapperSupportMaintainService。
+  - com.dwarfeng.fdr.stack.service.TriggerSupportMaintainService。
+  - com.dwarfeng.fdr.stack.service.WasherSupportMaintainService。
 
 ---